refactor: 사용자 도메인 패키지 구조 레이어드 스타일로 리팩토링
