+++ src/main/java/com/example/daobe/lounge/application/LoungeFacadeService.java
@@ -7,15 +7,15 @@ import com.example.daobe.lounge.application.dto.LoungeInfoDto;
 import com.example.daobe.lounge.application.dto.LoungeInviteDto;
 import com.example.daobe.lounge.domain.Lounge;
 import com.example.daobe.lounge.domain.LoungeResult;
+import com.example.daobe.lounge.domain.LoungeSharer;
+import com.example.daobe.lounge.domain.repository.LoungeSharerRepository;
 import com.example.daobe.lounge.exception.LoungeException;
 import com.example.daobe.lounge.exception.LoungeExceptionType;
 import com.example.daobe.objet.domain.repository.ObjetRepository;
-import com.example.daobe.lounge.domain.LoungeSharer;
-import com.example.daobe.lounge.domain.repository.LoungeSharerRepository;
-import com.example.daobe.user.entity.User;
+import com.example.daobe.user.domain.User;
+import com.example.daobe.user.domain.repository.UserRepository;
 import com.example.daobe.user.exception.UserException;
 import com.example.daobe.user.exception.UserExceptionType;
-import com.example.daobe.user.repository.UserRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
