+++ src/main/java/com/example/daobe/lounge/application/LoungeService.java
@@ -10,7 +10,7 @@ import com.example.daobe.lounge.domain.LoungeType;
 import com.example.daobe.lounge.domain.repository.LoungeRepository;
 import com.example.daobe.lounge.exception.LoungeException;
 import com.example.daobe.lounge.exception.LoungeExceptionType;
-import com.example.daobe.user.entity.User;
+import com.example.daobe.user.domain.User;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
