+++ src/main/java/com/example/daobe/objet/application/ObjetService.java
@@ -20,9 +20,9 @@ import com.example.daobe.objet.domain.ObjetType;
 import com.example.daobe.objet.domain.repository.ObjetRepository;
 import com.example.daobe.objet.domain.repository.ObjetSharerRepository;
 import com.example.daobe.objet.exception.ObjetException;
-import com.example.daobe.user.entity.User;
+import com.example.daobe.user.domain.User;
+import com.example.daobe.user.domain.repository.UserRepository;
 import com.example.daobe.user.exception.UserException;
-import com.example.daobe.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import java.util.HashSet;
 import java.util.List;
