+++ src/main/java/com/example/daobe/objet/domain/Objet.java
@@ -3,7 +3,7 @@ package com.example.daobe.objet.domain;
 import com.example.daobe.common.entity.BaseTimeEntity;
 import com.example.daobe.lounge.domain.Lounge;
 import com.example.daobe.objet.application.dto.ObjetCreateResponseDto;
-import com.example.daobe.user.entity.User;
+import com.example.daobe.user.domain.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
