+++ src/main/java/com/example/daobe/user/domain/DefaultNicknamePolicy.java
@@ -1,4 +1,4 @@
-package com.example.daobe.user.entity;
+package com.example.daobe.user.domain;
 
 import java.util.UUID;
